tools/requirements.txt
Flask
Flask-CORS
PyJWT
requests
pandas
openpyxl
qrcode[pil]
boto3
urllib3
